+++ dataContract_test.go
@@ -1,11 +1,15 @@
 package main
 
 import (
+	"fmt"
+	"os"
 	"reflect"
 	"testing"
 )
 
 func TestGetValue(t *testing.T) {
+	model, _ := os.ReadFile("./test_resources/model.yaml")
+
 	type args struct {
 		contract DataContract
 		path     []string
@@ -45,6 +49,26 @@ func TestGetValue(t *testing.T) {
 				path:     []string{"schema", "type"}},
 			wantErr: true,
 		},
+		{
+			name: "local reference",
+			args: args{
+				contract: DataContract{"schema": map[string]interface{}{
+					"specification": "$ref: test_resources/model.yaml",
+				}},
+				path: []string{"schema", "specification"}},
+			wantValue: string(model),
+			wantErr:   false,
+		},
+		{
+			name: "remote reference",
+			args: args{
+				contract: DataContract{"schema": map[string]interface{}{
+					"specification": fmt.Sprintf("$ref: %v/model.yaml", TestResourcesServer.URL),
+				}},
+				path: []string{"schema", "specification"}},
+			wantValue: string(model),
+			wantErr:   false,
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -55,6 +79,49 @@ func TestGetValue(t *testing.T) {
 			}
 			if !reflect.DeepEqual(gotValue, tt.wantValue) {
 				t.Errorf("GetValue() gotValue = %v, want %v", gotValue, tt.wantValue)
+			}
+		})
+	}
+}
+
+func TestGetDataContract(t *testing.T) {
+	dataContract := DataContract{
+		"dataContractSpecification": "0.9.0",
+		"id":                        "my-data-contract-id",
+		"info": map[string]interface{}{
+			"title":   "My Data Contract",
+			"version": "0.0.1",
+		},
+	}
+	type args struct {
+		location string
+	}
+	tests := []struct {
+		name                   string
+		args                   args
+		wantDataContractObject DataContract
+		wantErr                bool
+	}{
+		{
+			name:                   "local",
+			args:                   args{location: "test_resources/datacontract.yaml"},
+			wantDataContractObject: dataContract,
+		},
+		{
+			name:                   "remote",
+			args:                   args{location: fmt.Sprintf("%v/datacontract.yaml", TestResourcesServer.URL)},
+			wantDataContractObject: dataContract,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gotDataContractObject, err := GetDataContract(tt.args.location)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("GetDataContract() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(gotDataContractObject, tt.wantDataContractObject) {
+				t.Errorf("GetDataContract() gotDataContractObject = %v, want %v", gotDataContractObject, tt.wantDataContractObject)
 			}
 		})
 	}