+++ main.go
@@ -16,13 +16,31 @@ func main() {
 	fileNameFlag := &cli.StringFlag{
 		Name:  "file",
 		Value: dataContractFileName,
-		Usage: "file name for the data contract",
+		Usage: "location of the data contract, path or url (except init)",
+	}
+
+	schemaTypePathFlag := &cli.StringFlag{
+		Name:  "schema-type-path",
+		Value: "schema.type",
+		Usage: "definition of a custom path to the schema type in your data contract",
+	}
+
+	schemaSpecificationPathFlag := &cli.StringFlag{
+		Name:  "schema-specification-path",
+		Value: "schema.specification",
+		Usage: "definition of a custom path to the schema specification in your data contract",
+	}
+
+	withFlag := &cli.StringFlag{
+		Name:     "with",
+		Required: true,
+		Usage:    "location (url or path) of the stable version of the data contract",
 	}
 
 	app := &cli.App{
 		Name:    "datacontract",
 		Usage:   "Manage your data contracts 📄",
-		Version: "v0.2.1",
+		Version: "v0.3.0",
 		Authors: []*cli.Author{
 			{Name: "Stefan Negele", Email: "stefan.negele@innoq.com"},
 		},
@@ -92,6 +110,30 @@ func main() {
 				},
 			},
 			{
+				Name:  "schema",
+				Usage: "print schema of the data contract",
+				Flags: []cli.Flag{fileNameFlag, schemaSpecificationPathFlag},
+				Action: func(ctx *cli.Context) error {
+					pathToSpecification := strings.Split(ctx.String("schema-specification-path"), ".")
+					return PrintSchema(ctx.String("file"), pathToSpecification)
+				},
+			},
+			{
+				Name:  "quality",
+				Usage: "print quality checks of the data contract",
+				Flags: []cli.Flag{
+					fileNameFlag,
+					&cli.StringFlag{
+						Name:  "quality-specification-path",
+						Value: "quality.specification",
+						Usage: "definition of a custom path to the quality specification in your data contract",
+					}},
+				Action: func(ctx *cli.Context) error {
+					pathToSpecification := strings.Split(ctx.String("quality-specification-path"), ".")
+					return PrintQuality(ctx.String("file"), pathToSpecification)
+				},
+			},
+			{
 				Name:  "open",
 				Usage: "save and open the data contract in Data Contract Studio",
 				Flags: []cli.Flag{fileNameFlag},
@@ -103,21 +145,9 @@ func main() {
 				Usage: "EXPERIMENTAL (dbt specification only) - show differences of your local and a remote data contract",
 				Flags: []cli.Flag{
 					fileNameFlag,
-					&cli.StringFlag{
-						Name:     "with",
-						Required: true,
-						Usage:    "url of the stable version of the data contract",
-					},
-					&cli.StringFlag{
-						Name:  "schema-type-path",
-						Value: "schema.type",
-						Usage: "definition of a custom path to the schema type in your data contract",
-					},
-					&cli.StringFlag{
-						Name:  "schema-specification-path",
-						Value: "schema.specification",
-						Usage: "definition of a custom path to the schema specification in your data contract",
-					},
+					withFlag,
+					schemaTypePathFlag,
+					schemaSpecificationPathFlag,
 				},
 				Action: func(ctx *cli.Context) error {
 					pathToType := strings.Split(ctx.String("schema-type-path"), ".")
@@ -130,27 +160,22 @@ func main() {
 				Usage: "EXPERIMENTAL (dbt specification only) - detect breaking changes between your local and a remote data contract",
 				Flags: []cli.Flag{
 					fileNameFlag,
-					&cli.StringFlag{
-						Name:     "with",
-						Required: true,
-						Usage:    "url of the stable version of the data contract",
-					},
-					&cli.StringFlag{
-						Name:  "schema-type-path",
-						Value: "schema.type",
-						Usage: "definition of a custom path to the schema type in your data contract",
-					},
-					&cli.StringFlag{
-						Name:  "schema-specification-path",
-						Value: "schema.specification",
-						Usage: "definition of a custom path to the schema specification in your data contract",
-					},
+					withFlag,
+					schemaTypePathFlag,
+					schemaSpecificationPathFlag,
 				},
 				Action: func(ctx *cli.Context) error {
 					pathToType := strings.Split(ctx.String("schema-type-path"), ".")
 					pathToSpecification := strings.Split(ctx.String("schema-specification-path"), ".")
 
 					return Breaking(ctx.String("file"), ctx.String("with"), pathToType, pathToSpecification)
+				},
+			}, {
+				Name:  "inline",
+				Usage: "inline all references specified with '$ref' notation",
+				Flags: []cli.Flag{fileNameFlag},
+				Action: func(ctx *cli.Context) error {
+					return Inline(ctx.String("file"))
 				},
 			},
 		},