+++ quality.go
@@ -27,6 +27,20 @@ func qualityCheck(contract DataContract) error {
 	}
 
 	fmt.Println("Data set: %w", dataset)
+)
+
+func PrintQuality(dataContractLocation string, pathToQuality []string) error {
+	dataContract, err := GetDataContract(dataContractLocation)
+	if err != nil {
+		return fmt.Errorf("failed parsing local data contract: %w", err)
+	}
+
+	qualitySpecification, err := getQualitySpecification(dataContract, pathToQuality)
+	if err != nil {
+		return fmt.Errorf("can't get specification: %w", err)
+	}
+
+	fmt.Println(string(TakeStringOrMarshall(qualitySpecification)))
 
 	return nil
 }