+++ string.go
@@ -1,5 +1,7 @@
 package main
 
+import "net/url"
+
 func EqualStringPointers(s1, s2 *string) bool {
 	// both are the same (e.g. nil)
 	if s1 == s2 {
@@ -20,4 +22,19 @@ func StringPointerString(str *string) string {
 	}
 
 	return *str
+}
+
+func IsURI(reference string) bool {
+	_, err := url.ParseRequestURI(reference)
+	return err == nil
+}
+
+func TakeStringOrMarshall(object interface{}) []byte {
+	var bytes []byte
+	if str, isString := object.(string); isString {
+		bytes = []byte(str)
+	} else if mp, isMap := object.(map[string]interface{}); isMap {
+		bytes, _ = ToYaml(mp)
+	}
+	return bytes
 }